feature(16): Refactor call to api and remove all unnesesary css
