+++ src/app/components/app/app.component.ts
@@ -9,6 +9,5 @@ import { MatListModule } from '@angular/material/list';
   standalone: true,
   imports: [CommonModule, RouterOutlet, HeaderComponent, MatListModule],
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {}
