+++ src/app/core/services/auth-header/auth-header.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AuthHeaderService } from './auth-header.service';
+
+describe('AuthHeaderService', () => {
+  let service: AuthHeaderService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(AuthHeaderService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
