+++ src/app/includes/button/components/root/button.component.ts
@@ -6,7 +6,6 @@ import { NgClass } from '@angular/common';
   standalone: true,
   imports: [NgClass],
   templateUrl: './button.component.html',
-  styleUrl: './button.component.scss',
 })
 export class ButtonComponent {
   @Input() classes?: string;
