+++ src/app/includes/error-loading-data/components/root/error-loading-data.component.ts
@@ -5,7 +5,6 @@ import { Component, Input } from '@angular/core';
   standalone: true,
   imports: [],
   templateUrl: './error-loading-data.component.html',
-  styleUrl: './error-loading-data.component.scss',
 })
 export class ErrorLoadingDataComponent {
   @Input() message: string = '';
