--- src/app/includes/product-tile/components/product-tile-image/product-tile-image.component.spec.ts
@@ -1,22 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ProductTileImageComponent } from './product-tile-image.component';
-
-describe('ProductTileImageComponent', () => {
-  let component: ProductTileImageComponent;
-  let fixture: ComponentFixture<ProductTileImageComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ProductTileImageComponent],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(ProductTileImageComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
