+++ src/app/includes/product-tile/components/product-tile-name/product-tile-name.component.ts
@@ -5,7 +5,6 @@ import { Component, Input } from '@angular/core';
   standalone: true,
   imports: [],
   templateUrl: './product-tile-name.component.html',
-  styleUrl: './product-tile-name.component.scss',
 })
 export class ProductTileNameComponent {
   @Input() productName?: string;
