+++ src/app/includes/product-tile/components/root/product-tile.component.ts
@@ -18,7 +18,6 @@ import { ProductTileImageComponent } from '@includes/product-tile/components/pro
     ProductTileImageComponent,
   ],
   templateUrl: './product-tile.component.html',
-  styleUrl: './product-tile.component.scss',
 })
 export class ProductTileComponent {
   @Input() product?: Product;
