+++ src/app/pages/not-found/components/root/not-found.component.ts
@@ -5,6 +5,5 @@ import { Component } from '@angular/core';
   standalone: true,
   imports: [],
   templateUrl: './not-found.component.html',
-  styleUrl: './not-found.component.scss',
 })
 export class NotFoundComponent {}
