+++ src/app/pages/shop/components/categories/core/services/category.service.ts
@@ -1,30 +1,17 @@
 import { inject, Injectable } from '@angular/core';
 import { injectQuery } from '@ngneat/query';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { Category } from '@pages/shop/components/categories/core/types/category/category';
 import { environment } from '@environments/environment';
+import { AuthHeaderService } from '@core/services/auth-header/auth-header.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class CategoryService {
   private readonly httpClient = inject(HttpClient);
+  private readonly authHeaderService = inject(AuthHeaderService);
   private readonly query = injectQuery();
-  private readonly authHeader = this.createBasicAuthHeader(
-    environment.WP_API_USERNAME,
-    environment.WP_API_PASSWORD,
-  );
-  constructor() {}
-
-  private createBasicAuthHeader(
-    username: string,
-    password: string,
-  ): HttpHeaders {
-    const encodedCredentials = btoa(`${username}:${password}`);
-    return new HttpHeaders({
-      Authorization: `Basic ${encodedCredentials}`,
-    });
-  }
 
   getCategories() {
     return this.query({
@@ -32,7 +19,7 @@ export class CategoryService {
       queryFn: () => {
         return this.httpClient.get<Category[]>(
           `${environment.BASE_WP_API_URL}/products/categories`,
-          { headers: this.authHeader },
+          { headers: this.authHeaderService.authHeader },
         );
       },
     });
